Significant progress on the CLI overhaul
- The primary entrypoint is now __main__.py. This enables usage of adles with "python -m adles", and makes adding and running scripts easier
- Scripts are now called as subcommands instead of individual .py files. This simplifies a *lot* of things, such as importing, duplication, arguments and setup, etc.
- Vsphere scripts have their own subclass
- Improved the dunder methods and ability to get help/version info
- Moved script-related utils into scripts/script_utils.py
- Consolidated all of the individual vSphere scripts as classes in scripts/vsphere_scripts.py
